Feature: Add option to filter profiles by archive type for `xcode-project use-profiles`  (#224)
