+++ docs/xcode-project/use-profiles.md
@@ -13,6 +13,7 @@ xcode-project use-profiles [-h] [--log-stream STREAM] [--no-color] [--version] [
     [--custom-export-options CUSTOM_EXPORT_OPTIONS]
     [--warn-only]
     [--code-signing-setup-verbose-logging]
+    [--archive-method USE_PROFILE_ARCHIVE_METHOD]
 ```
 ### Optional arguments for action `use-profiles`
 
@@ -40,6 +41,10 @@ Show warning when profiles cannot be applied to any of the Xcode projects instea
 
 
 Show verbose log output when configuring code signing settings for Xcode project. If not given, the value will be checked from the environment variable `XCODE_PROJECT_CODE_SIGNING_SETUP_VERBOSE_LOGGING`.
+##### `--archive-method=ad-hoc | app-store | development | enterprise`
+
+
+Use only the profiles that are eligible for given archive method for code signing setup. If not specified, all found profiles will be used.
 ### Common options
 
 ##### `-h, --help`
